tests/wpt/web-platform-tests/app-history/navigate/goTo-multiple-steps.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(appHistory.entries().length, 1);
  let key0 = appHistory.current.key;
  await appHistory.navigate("#1").committed;
  await appHistory.navigate("#2").committed;
  let key2 = appHistory.current.key;
  assert_equals(appHistory.entries().length, 3);

  await appHistory.goTo(key0).committed;
  assert_equals(appHistory.entries().length, 3);
  assert_equals(appHistory.current, appHistory.entries()[0]);
  assert_equals(key0, appHistory.current.key);
  await appHistory.goTo(key2).committed;
  assert_equals(appHistory.entries().length, 3);
  assert_equals(appHistory.current, appHistory.entries()[2]);
  assert_equals(key2, appHistory.current.key);
}, "goto() can precisely traverse multiple steps in the joint session history");
</script>
